remove the code for videos
