+++ src/Annotator/reducers/get-active-image.js
@@ -4,7 +4,6 @@ export default (state) => {
   let currentImageIndex = null,
     pathToActiveImage,
     activeImage
-  if (state.annotationType === "image") {
     currentImageIndex = state.selectedImage
     if (currentImageIndex === -1) {
       currentImageIndex = null
@@ -13,9 +12,5 @@ export default (state) => {
       pathToActiveImage = ["images", currentImageIndex]
       activeImage = getIn(state, pathToActiveImage)
     }
-  } else if (state.annotationType === "video") {
-    pathToActiveImage = ["keyframes", state.currentVideoTime || 0]
-    activeImage = getIn(state, pathToActiveImage) || null
-  }
   return { currentImageIndex, pathToActiveImage, activeImage }
 }
