+++ src/Annotator/reducers/image-reducer.js
@@ -6,7 +6,7 @@ export default (state, action) => {
   const {currentImageIndex} =
     getActiveImage(state)
 
-  if (action.type === "IMAGE_OR_VIDEO_LOADED") {
+  if (action.type === "IMAGE_LOADED") {
     return setIn(state, ["images", currentImageIndex, "pixelSize"], {
       w: action.metadata.naturalWidth,
       h: action.metadata.naturalHeight,
