+++ src/Localization/translation-de-DE.js
@@ -3,7 +3,6 @@ const translationDeDE = {
   "menu.history": "Verlauf",
   "menu.regions": "Regionen",
   "error.image": "Bild konnte nicht geladen werden.",
-  "error.imagevideo": "Fehler im Bild oder Video",
   "no.history": "Kein Verlauf vorhanden",
   "desc.class": "Klassen",
   "move.polypoint": "Polygon Punkt verschoben",
